builder: changed how code is written for handling continuation of routines
For routine trials there was a variable trialsContinue etc. and
there was a test "while trialsContinue:" but this was causing errors
with copied/pasted routines. Simpler to have a variable `continueRoutine`
that gets updated at start of each routine to point to the current one.
That way components don't need to know which routine they belong to.

Signed-off-by: Jonathan Peirce <jon@peirce.org.uk>