* buildingOnAMac.txt: Use a txt file instead of a richtext file.
Notes about installation on a mac should be kept in an easily-edited
plain text file rather than in a rich text file.  There's no
formatting in the document that requires the rich text format.

This text file has some fancy RST-ish section markers, but that can be
removed if desired.