builder: fullscr window size and warning msg
Issue: At the end of a builder script, I would always get a warning about
a different screen size [1440, 990] from the requested [1024,768]. That is, when
writing out a script it was not using my real size [1440, 990], despite displaying
it correctly in settings. And my monitor is set for this size in MonitorCenter.

Fix: Obtain size info from wx.Display() just before writing out a script, ignoring
the value in self.params['Window size (pixels)'] if full screen is requester.
Another approach would be to correct that param value somehow.