fullRandom option for TrialHandler: randomize list order completely
Example:
1-6 repeated five times in fullRandom order, 5 4 1 1 3 6 2 2 1 3 ...
so there can be runs of the same trial type (up to nReps in a row),
and not all items need to have been presented before an item can
repeat (unlike in 'random').

For more complicated sequences, users can generate any order they like in excel,
and run in 'sequential' order.

Not sure about (did not try to address):
- letting stair or multistair use fullRandom as an option