Unicode for saveAsWideText function
Added UTF-8 encoding to the TrialHandler's saveAsWideText function.
In response to mailing list request by Robert Busching, 26 Mar 2012.

NB there is now some duplication between the same-named function in
both the TrialHandler and the ExperimentHandler classes.  Inevitably,
duplication leads to inconsistency problems like this.  Also the
appendFile parameter defaults differently in each version of the
function, which might cause problems.
saveAsWideText() needs to stay as a function of the TrialHandler for
those people who may be using it in the Coder, and who aren't using the
ExperimentHandler class, but the function is now just a special case of
the (more capable) function in the Experiment Handler.  Can the two be
reconciled so that code duplication is minimised?