ENH : window.flip now returns flip time
If waitBlanking is True during the current call to win.flip(), have flip
return the time that was read immediately following the GL.glFinish() in
the waitBlanking blocking code.

This

a) makes for cleaner user script code when someone waht to get the time
of the flip (for calculating an RT start time for example) and

b) may be somewhat more accurate than having someone call the get time
function after flip returns, given there is a small amount of extra
logic preformed in flip() between the end of the blocking code the and
function return.