Add explanation of Expected Duration caveat to non-slip timing
When I first used it, I expected a component's (and therefore a routine's) expected duration to be sort of a suggestion used to draw the timeline/flow, etc. However, when I time it and look, it appears to be more like a "max" duration due to the non-slip timing.

This Note points out that caveat.