BF #157 Gracefully handle an unavailable screen
If a second screen was requested, but only one was available, both the Builder script writer and visual.Window would crash when trying to run. 

This fix allows both the Builder and Window to warn about the screen and then graciously step down to the available screen, instead of throwing an error. Builder will temporarily write the screen as 0 if the requested one can't be found, and Window itself which will run on any available screen.

It's probably alright for this to be an Error that halts, but I think warning and running on the available screen is probably a better choice for most people. If we really want the script to fail to build when a screen's not available, perhaps we could add some check to see if Debug mode has been enabled, but I think it's probably better this way.