BF: removed the incorrect scaling that was being applied to LMS color calibration
The use of scaling to the whitepoint of the monitor may be reinstated as an
option later, but can only be done as the rgb values are computed, not as part
of the conversion matrix itself.

A warning has been added for anyone trying to use LMS space in a stimulus
if their calibration was performed during the time that the error was in
place (version <1.76) although the number of users that this will affect
is expected to be extremely small.