ENH: added the basic hooks for an ExpFile object in Builder Experiments
The ExpFile represents a psyexp file on disk and simply writes the code
represented by that experiment's Flow in the appropriate places in its
own experiment code.

As yet, this is untested and we need to work out how to specify filename
from the GUI and when to read in the file (e.g. should we reload from disk each
time we're asked to generate code?)