Refactor parallel port access to use classes instead of globals.
Move different methods into their own classes which carry their
own state.  In order to maintain the existing API, we now
init the variable in psychopy/parallel.py as an object instead
of just a base address.  This allows us to abstract out which
platform we're on.  It also helps because we can pass references
to the parallel port handling code to other routines as necessary
and they don't need to care about how the port is being accessed.
Direct class usage also allows easy handling of multiple parallel
ports.

Signed-off-by: Mark Hymers <mark.hymers@ynic.york.ac.uk>