BF: fixes to quick fix issues found on Linux/OS X
- All examples (not including eye tracker examples folder, not yet
tested) run on all 3 platforms.
- Mouse button bug on Linux found and reported to issue tracker, making
the ioMouse example run, but not due what it should.
- OSX OK, other than the ioHubServer having to be killed every time the
program ends; must be somemthing related to an OSX specific API being
used, as this does not happen on other OS's. So pyobjc related APIs
being used will be a good place to start looking.