ENH: iohub docs and new example
- added 125 line extended_iohub/datastore/events2txt.py example. It
allows someone to pick a DataStore HDF5 file using a file browser, then
select which event type they want to save to a tab delimited file using
a drop down list dialog that only includes event types for which there
is data for in the selected file. All the events of the selected type
are streamed to a .txt file using the event iterator so memory is kept
low regardless of file size. Speed depends on the event type: Message
events are saved at about 50,000 events / sec, mouse events at about
30,000 events / sec on my i7 PC for example.
- added the iohub docs to the top level dir in a folder called
iohub_docs. These have been changed to indicate the iohub merge with
psychopy and all the module paths now use psychopy.iohub. These are
interm docs until proper doc integration can be done in pscyhopy docs;
but something needs to be available until then and this is what I got.
;)  I need to have them with the iohub source so as I work on
psychopy.iohub I can keep the 'current' iohub docs updated.
- I'm not expecting that the built docs be 'hosted' the same place that
the pscyhpy docs are (unless you want to, then that would be great); I
can still have them available online using the current URL.  If you know
of a free place that could host them so that people do not need to go to
isolver-solutions.com to see the docs, please let me know, as that would
be great.