TST: better testing of script compile (routine needing adding to the flow)
previously components were created and added to the routine, but because
that wasn't on the flow it was never having its code written. That's now
fixed and should increase coverage.

Along the way I had to make some changes to parameters. These wouldn't have
mattered for real-world Builder, because they would have been converted
to strings by the wx dialog.