ENH: Extending TrialHandler's n-back information options
In response to user query at
<https://groups.google.com/forum/#!topic/psychopy-users/9aeU1AgEJv4>.
Sometimes access is needed to information from previous trials (as in
an n-back task). As a kludge, the trialHandler's getFutureTrial()
function could be used for this by passing a negative index value, but
as it wasn't designed for this, it didn't include a bounds check to
handle trying to access a trial before the first one. This has now been
added.
To be explicit, have also added a getEarlierTrial() function, which
simply calls the getFutureTrial function but ensures the index is
negative.