DOC: Fix for namespace issues in psychopy.data API examples
A user found problems when trying to create a Multistairhandler:
<https://groups.google.com/forum/#!topic/psychopy-users/kPEtdAp5C_s>

This was due to the docs not specifying that handler classes are
instantiated with a "data" namespace prefix. Fixed this to make all of
the examples in the comments consistent (previously the correct syntax
was given only for one QuestHandler example).