ENH: basic hooks and standalone dependencies to run ipython notebooks
The notebook works with psychopy, and can generate windows, collect
responses etc

BUT it can't be shut down and wx ceases to respond, so the app has to
be reset! So it's currently disabled, pending someone spawning it from
a thread or process. But the dependencies (tornado and pyzmq) are now
included