BF: Improved labjack data handling
- If data is dropped by labjack USB HID interface, iohub labjack device
records proper amount of missing data and continues with reporting.
- If chosen sampling rate results in USB HID packets being returned by
labjack that have a first or last sample being a partial read of the N
channels being scanned, the partial channel sample is stored and the
remaining channel data for that sample is appended before the sample is
saved / streamed.