ENH: Send shutdown msg to ioHub on IDE Stop cmd
Added 'terminateHubProcess' method to the psychopy2 app class.

Called by stopFile() method of builder and coder.

This opens a udp socket to the iohub server address and sends the
STOP_IOHUB_SERVER msg. No reply is expected. udp operation times out in
1 second is msg is not sent.

If iohub is not running, an exception is caught and silently passed, so
it is fine to call anytime stopFile() is called.

This