ENH: TextBox can report if all text displayed.
If the text provided to a textbox is too long to be fully displayed
within the textbox region (after soft wrapping is applied), the text is
truncated when displayed.

User can call textbox.getDisplayedText() to get what text was actually
displayed and compare it to the text set and see if there is a diff.