BF: Fix KeyError when starting Builder experiment
The experiment crashed when starting from Builder and "Experiment info"
contained no field with name "participant", "Participant", "Subject" or
"Observer" but fields with names containing these names. For example if
a field was named "ParticipantID" this showed the dialog asking for the
value but crashed afterwards because "Participant" was looked up. Fault
was that the settings component checked the keys for experiment info by
looking into the serialized form of the dict and found sub strings this
way. This was fixed by using the eval()ed dict to check the keys there.