BF: ElementArrayStim initialization
Using a list or tuple for the xys argument when creating an
ElementArrayStim causes the setWithOperation function to crash. This was
due to the fact that the xys argument is not converted to a numpy array
at initialization while it is used in the setWithOperation function.
This was fixed by using the val2array function from arraytools for the
xys argument of the ElementArrayStim at initialization.