FF: Fix Aperture initialization
Passing a tuple, a list or an array did not work because checking the
type was done after checking if it was string. This could not be true
and therefore the default configuration was used. Passing None did work
accidentally because it sets the same configuration as the default case
but actually the related check also failed for the same reason.