BF: WindowsError fix that should work on other OS
This fix for the memory access failed exception should not give issues
on Linux or OS X.

BTW, while debugging I printed the lib that was causing the WindowsError
originally, and I got:

Unexpected error loading library gdk-x11-2.0: [Error 126] The specified
module could not be found.

No idea why it was trying to load  gdk-x11-2.0, but is fails because it
does not exist from what I can tell.