ENH: Fix ImportError if old wxpython is installed
Builder imports modules from wxpython 2.8 which are not available for
older versions. If the system had older versions installed and one of
these versions was set as the default than starting Builder raised an
ImportError. This was fixed by just using wxversion to ensure that at
least wxpython 2.8 is available which provides the needed agw module.