ENH: ioHub ioSync MCU device update
- analog input events work, demo added
- digital output requests work; demo added
- psychopy and ioSync time bases syncing to +/- 1 msec max error, demo
added
- AnalogInputDeviceDelayTest.py demo completed; used digital out
functionality of ioSync to test the timing error between when a analog
input event actually occurs and the labjack AnalogInput event times
being given. The error can then be applied as a setting to the LabJack,
resulting in timing with +/- 0.5 msec accuracy. Note: test must be run
on same computer using same LabJack as what will be used during data
collection.