BF and DOC : StairHandler using wrong stepSizes
BF : StairHandler was using the wrong stepSizes after a reversal : it applied the intensity increase/decrease BEFORE changing the step. This always made the first intensity change (after a reversal) use the former step.

DOC : updated comments about the StairHandler to include the fact that nUp and nDown are always considered as 1 until the first reversal is reached.