BF : Unicode warnings and errors in microphone
I've found what the Unicode problem was (I believe).

devString is already encoded by the OS character encoding, so trying to encode it again results in an error. We must first decode it, and then re-encode it in utf-8 to compare with the utf-8 encoded prefDriver. The problem was to make the code compatible with multiple versions of Windows that may not be using the same character encoding.

sys.getfilesystemencoding() should do the trick, but I need someone to test it at least on an english windows version, and it would be great to test it on others too. On my french version, it's working fine and no warnings are being thrown anymore.

Plus I added a .lower() to make it non case-sensitive.