ADD: Graphics and control logic for ET validation
A Validation wrapper is not yet done, but the core classes that will be
used to support an eye tracker independent validation process is now in
place.

Could also use same code for eye tracker calibration on systems that use
psychopy graphics for calibration stim.

See the 'validation.py' demo in demos.coder.iohub.eyetracker for an
example
of how the new classes will be used to create a packaged validation
process method call for the iohub eye tracker device.

TODO:
- Calculate and display validation accuracy
- Add starting intro / menu screen and validation complete type msg
screen.