BF: fixed scaling issue with dotstim. Fixes #554
Since the refactoring this was applying the fieldSize variable twice. 
In the new scheme the base vertices should always only be -0.5<xy<0.5
and leave the scaling to field size to occur later.

Added a test to make sure that pos and field size respect units
in future (using a dense array of dots as a test image)