ENH: Added USB Keyboard mode support to ioSync
There is now a python method in the iosync interface that allows your
program to have the iosync hardware generate a keyboard press and
release event which are sent to the computer via the USB cable.

The duration of the keypress can be set as a param in the command
method.

The keypress event occurs in realtime with < 1 msec of uncertainty about
the time that the HID keyboard event packet was sent to the OS via USB.

To use this new feature, the ioSync.ino teensy 3 program must be
compiled with Serial set to Serial + Keyboard.

When the ioSync is connected to the computer, it will appear as a
generic USB keyboard to the OS ( it should be listed in the keyboard
section of your device manager).

The idea is that this should be able to be used to look at the software
only differences between OSs and libraries being used, as the key event
time is know precisely and in the same time base  used for key events
received by psychopy or iohub.