BF: fix bug introduced to getKeys/waitKeys
CHange made in
https://github.com/psychopy/psychopy/commit/788292b736aa8d357ca5487d69ee02b3995ba79f
was causing issues for these functions when a clock instance was padded
in as the timer arg.

Also changed _onPygletText to use psychopy.clock.getTime()  instead of
psychopy.core.getTime() as it was before.

This fixes this specific (and big) bug I introduced, but does not
address other issues (which I thought the original commit was fixing).
Will look into this more later.