ENH: Fix compatibility with wxPython 2.8
Starting the Builder with wyPython 2.8 raised AttributeErrors because
Font's Larger() method is not available before 2.9.1. The application
calls ensureMinimal() for 2.8 so this version should be supported. It
was fixed by adding a check if Larger() is available. Using a smaller
font still seems to work and allows using PsychoPy on older wxPython.