DOC: Add an XSD for PSYEXP XML format
The XSD is only basic and does not care about most of the special cases.
For example there is no check for LoopInitiator and LoopTerminator pairs
and if Routines referenced in Flow are actually definied. But structures
and most of the other restrictions (unique names, enumerated valType and
updates, available Components) should be checked.