BF: fix to the time-stamped method of getKeys/waitKeys
This follows up on a3dfddc62 and 788292b736 which both arise from the
original move of clocks out of core and into a separate package. Closes
#653

For the call getKeys(timeStamped=True):

We really **should** use `core.getTime()` not `clock.getTime()` because
that uses time from creation not time since UTC.

For the call getKeys(timeStamped=customClock):

We need to find the **difference** in time bases between
`core.getTime()` and customClock.getTime() because the time base of
`core.getTime()` has already been substracted. This is the part that had
not been correct before and the attempt to get around it was breaking
the timeStamped=True version method of calling.