DOC: Remove tooltip non-functional suggestion
The tooltip previously gave an example of using 'randint(5)' as an argument to the row selection field in the loop dialog. This would actually lead to an error:
TypeError: range() integer end argument expected, got numpy.float64.
numpy.randint() does return an integer so not sure why this error arises, but removing the suggestion from the tooltip in the interim.