BF: typo in eyelink device class +
worked on trying to get rid of the error seen on win64 when wx3  + iohub
is imported:

runtime error R6031
- Attempt to initialize the CRT more than once.
This indicates a bug in your application.

Online posts suggest it could be from the windows LLHook for kb or mouse
events not being closed. Not sure that is the issue here though.