ENH: Ability to save xlib event deatils on linux os
Can now tell iohub to save the xlib event details that are used to
generate iohub kb and mouse events to a log file. This option
is set in the iohub/default_config.yaml:

# If True, OS level kb and mouse event details that iohub uses to
# generate associated device events will be logged. Only supported by
# linux right now. File is saved to experiment script folder,
# with name x11_events_{0}.log,
# where {0} = datetime.datetime.now().strftime("%Y_%m_%d_%H_%M")
log_raw_kb_mouse_events: False

Idea (meaning not yet implemented) is that this can be enabled to create
kb and mouse event sequences that can later be used to feed fake events
to iohub for keyboard and mouse event creation test cases.