BF: win64 and 32 bit freetype.dll files in path
Now tries to load freetype.dll on windows; if this fails (for example
because the freetype.dll found is a 64 bit dll and the python env is 32
bit), then freetype32.dll and freetype64.dll names are also tried.

Standalone Psychopy installer should be updated by renaming freetype.dll
to freetype32.dll in the psychopy2 DLLs folder.