BF: fixes to Bits++ (rendering glitch) and Bits# (not always finding identity LUT)
Main issue with Bits++ mode seems to be that the LUT was drawn twice
in succession and that confused the the Bits++ box.

Issue with Bits# not finding the LUT (only when bits++ mode requested)
was that it was drawing the bits++ header row and using that to do
the pixel testing. Now using row=5 to avoid any header rows.

Added some extra code to help with debugging findIdentityLUT()