ENH: iohub 'clearEvents' and proc priority setting
1) iohub clearEvents() had a 'race condition' (for lack of a better
term), causing native events that were time stamped by iohub, but not
yet fully converted into iohub events, not to be cleared when the
clearEvents() method was called.

This could result in very small negative time differences between the
time clearEvents() was called and occational event time stamps.

This issue would not occur if io.clearEvents('all') was used.

Now, any new native events that are 'in process', are fully processed
before the clearing logic is done. This fixes the issue / removes the
'race condition'.

2) iohub process priority API has been simplified. xxxbleHighPriority
and xxxRealtimePriority methods have been deprecated, in favour of a
single set and getPriority(level) method for the local process and the
iohub process.

See Computer.setPriority / getPriority and iohubConnection.setPriority /
getPriority methods for further info.