extraInfo was not being exported in wideText
The names of the extraInfo columns were being printed in the header, but the content wasn't! This is trialHandler only that I fixed, experimentHandler has the same bug which I can fix later, as well as having experimentHAndler also return dataframe, like trialHandler does. 
To do the fix, I broke up the header into extraInfo part and normal part. Otherwise, the code wouldn't work if extraInfo had a field name that was the same as the trials data (seems to be nothing stopping the user from doing that)