BF: Fix useVersion() not messaging actual errors
useVersion() raised ScriptError with message but there is no ScriptError
in Python and NameError was raised with a default message instead. This
made the raise only useful to exit the function but not to message back
what was wrong. It was fixed by replacing ScriptError with RuntimeError.

Found by pylint.