BF: code startVal should not be nonSlipSafe; hopefully closes #792
This change definitely addresses issue #792; it completely fixes both the minimal example, and the original
post to the users list.

I think this change is conceptually right (i.e., code for a start value means that start times are not always known, and hence are not safe to treat as non-slip friendly).

But I'm not certain that *everything* is right. Please review (as always).