BF: Replace mutable default arguments
Lists are mutable; using them as a default
method parameters can have undesired side-effects.
This fix replaces the default values of `stimOut`
in the method signatures with `None`; inside
the method body, a test is then performed which
sets `stimOut` to an empty list if no other value
was supplied.