BF: Don't disable auto-saving on manual save
ExperimentHandler.saveAsPickle() and
ExperimentHandler.saveAsWideText() used to disable
automatic saving once called, even if the user explicitly
requested auto-save on object creation time. This could
lead to undesired behavior or data loss.

Since we handle output filename collission now, there is no
need for this behavior in the case of saveAsWideText() anymore.

As for saveAsPickle(), Jon pointed out [1] that:

  * self.savePickle should be set to False just before
    dumping the pickle

  * the pickle be dumped

  * self.savePickle be restored to its initial value

The reason for that is that if a user loads a pickled (.psyexp)
file, they will most likely not want PsychoPy to auto-save the
pickle again. This fix takes care of that.

References:
[1] https://groups.google.com/d/msg/psychopy-dev/Z4m_UX88q8U/UGuh1eeyjMEJ