Update sound.rst: Volume parameter missing
I just saw that the documentation for the "Volume" parameter is missing. I was guessing (and there is a tool tip) what it is supposed to mean. I just wasn't sure whether, e.g., the value can be outside the range and will be clipped etc. Maybe someone can improve on this.

Best,

Axel