ENH: always draw ISI components in routine panel; closes #802
As noticed on the user list, having a $code duration or start means that an ISI (static) component
is not drawn at all in the routine panel. This made it awkward (imposssible?) to edit or delete it.

Fix: Ensure we always draw something. Append '???' to the name to signal that its timing is unknown / nonliteral.