BF: stimuli sometimes don't appear after a texture has been created
In some unusual circumstances stimuli could be made to disappear (if a
texture had been created and then a stimulus that did not have a texture
was drawn before any textured stimulus!)

Explanation: _createTexture was leaving the texture bound at the end of the function and if the final pixel was transparent that affected drawing operations. This was being reverted by the draw() function of any stimulus with a texture (that did unbind all textures) so wasn't typically seen.

I also altered the import function for pyglet.gl to be nicer with name
space