BF: Tobii calibration window not closing in 1.82
For some reason, in 1.82, when iohub calls win.close() on the full
screen window created for tobii calibration graphics, the window does
not actually close.

Work around is to minimize() the full screen window before calling
close, then it closes the window fine and the experiment window is
visible as expected.