ENH/BF: A function to set mouse exclusivity with Pyglet (fixes unwanted behavior in multi-monitor settings)
In a multi-monitor setting, the cursor can drift and thereby psychopy might not get the events from the experiment monitor. A function, setMouseExclusivity(), works with Pyglet to bind the mouse to the experiment window.

The price paid for using it is that 1. The cursor vanishes, 2. Only relative positions have meaning, i.e. getPos() returns [0, 0]. A proper warning has been put in place as I believe there are people who need this in special cases (myself included).