BF: Fix locking out of Builder Experiment Settings
If the "Screen" setting in the "Experiment Settings" dialog was set to
an unsupported text value, the dialog could not be opened again after
closing it. Values which could not be converted to integers raised a
ValueError in int() and values which could be converted but were smaller
1 raised a ValueError in wx.Display(). Both cases were fixed by adding
additional checks and using the default value (first screen) if needed.

The change also fixed not updating the "Window size" field if the window
setting was toggled while an unsupported "Screen" value was set above.