BF: Make StairHandler work with stepSizes other than float, int
StairHandler.__init__() used to check if the `stepSizes` argument was a
float or an int, and treated it as array-like otherwise. However, this would
fail e.g. for NumPy types like np.float32.

We now try to treat `stepSizes` as array-like without prior checking, and catch
exceptions if that fails.