Added Pen Pos. Validation
Also fixed some bugs. For accell dividebyzero error, it is because some
samples received from the wintab driver on the laptops have the same
hardware timestamp, but different data.

For now, I just catch the error and print out the event id's at issue.
vel and accell are set to 0. Need to look into best way to work around
this issue before making a 'proper' fix.