BF/RF: Ensure all MultiStairHandler staircases "finish"
Check if `QuestHandler` is finished immediately after adding a response, or only one
QUEST staircase will finish of all those executed through a `MultiStairHandler`.

Very much simplify `MultiStairHandler._createStairs()`: Instead of explicitly defining
paramters for the to-be-invoked staircases here, rely on the defaults of the regarding
StairHandler/QuestHandler classes. This change is backward-compatible.

Small stylistic changes.

Closes #1015.