ENH: Added first version of pysoundcard support
pysoundcard and pysoundfile are very thin wrappers around portaudio
and ffmpeg, respectively. The resulting modules are pretty simple
(compared to pyo for instance) and I hope we can be in better control
as a result. We don't need to create any additional process to run
pysoundcard so hopefully we'll also have less hassle with memory too.

I've set this up to use the callback system for writing chunks to the
audio buffer. That seems OK and we can control the buffer size on a
per-sound basis (I'm not sure what the optimal vals are).

Done:
  - looping
  - numpy arrays and tones

Not checked:
  - timing against visual/oscilloscope
  - audio files
  - stereo/mono streams

Not started:
  - volume control (should be simple multiplier in callback)
  - fadeout (callback needs to track time since fadeout was called?)

( @jeremygray @danielshub may be interested in this )