BF OSError: gamma.py can't find Carbon framework
gamma.py looks for the Carbon framework in the wrong place under OS X 10.11. The framework binary was moved from
> /System/Library/Carbon.framework/Carbon
to
> /System/Library/Frameworks/Carbon.framework/Carbon

Additionally gamma.py only calls the CGSetDisplayTransferByTable(,,,,). This function was moved to the CoreGraphics framework.

Therefore, my suggestion is that gamma.py should try to load the CoreGraphics framework first and should use the Carbon framework only as fallback.