ENH: Added rename Routine menu item
Has functions that change name of routine to a user specified name.
Changes notebook, namespace and routines dictionary.

ISSUE
Will not compile script if component added after renaming.  This is
despite there being no reference in the sublist or any dictionary.
Interestingly, if the name is changed for the second time with the same
name - makeValid is invoked and e.g. fooBar_2 is returned.  There is no
visible reference to fooBar in self.user.

The error message returned when script compilation fails is:
File "/psychopy/psychopy/app/builder/components/_base.py", line 326, in
getPosInRoutine
    routine = self.exp.routines[self.parentName]
KeyError: 'hello_world'

I have also included a context menu entry but this is not active yet.