NF: SimpleRatingScale
The SimpleRatingScale is a new type of (continuous) rating scale that can be
displayed both horizontally and vertically. It can therefore be used to create
visual analog scales (VAS) and labeled magnitude scales (LMS).

The initial mouse cursor position when first displaying the scale can be set
via the `mousePos` keyword argument.

This class supports ioHub for response time acquisition. When `iohub=True`
is passed to `__init__()`, we will try to connect to an existing ioHub session and
 acquire a mouse device. Since ioHub button press events will in some cases
 (e.g., when not running in fullscreen mode) return the mouse position in
 absolute display pixels without relation to the actual PsychoPy window, we
 use `event.Mouse.isPressedIn()` to check whether a mouse click within a valid
  area was registered, and only subsequently check the ioHub event queue to
  fetch that event (and its respective response time).

Since `_init__()` tended to become too large and cluttered, many initialization
procedures have been placed in separate functions (`_gen_*()`), as commonly
done by other stimulus classes in PsychoPy. However, I avoided the "silent"
creation of attributes in these functions: they are all side-effect free, and return
values that are then explicitly assigned to an attribute inside `__init__()`. So by
just looking at `__init__()`, it is already clear which attributes the class has, and
when and how they are being created.

Further, all internal implementation details follow, to the best of my knowledge,
the PEP-8 style and naming recommendations.

The public API is very minimalistic, as, aside from the provided keyword arguments,
only the following attributes and methods are exposed:

    - `draw()`
    - `waitForResponse`
    - `reset()`
    - `response`
    - `rt`

 Currently, only a small number of attributes is user-settable after instantiation:

    - `mousePos`
    - `maxTime`
    - `finishOnResponse`
    - `resetOnFirstFlip`
    - `autoLog`

 If `resetOnFirstFlip=True` (the default), the `reset()` method will be called immediately
 upon the next `visual.Window`.flip()` (via `Window.callOnFlip()`). This ensures that
 response times will be precisely time-locked to the onset of the rating scale presentation.

 # Please enter the commit message for your changes. Lines starting